Refactor `FunctionDemo` component
- split out every component into a separate file
- clean-up the code
- add comments for the parts that are not clear at the first glance
- refactor `FunctionSignature` component
- renamed `FunctionConfigurator` into `Configurator`
