+++ apps/native-component-list/src/components/FunctionDemo/index.ts
@@ -1,0 +1,5 @@
+import FunctionDemo, { FunctionDescription } from './FunctionDemo';
+export default FunctionDemo;
+
+export { FunctionDescription };
+export { FunctionParameter } from './index.types';
